# lib/api-zod/src/generated/types/analyticsData.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartDataPoint } from './chartDataPoint';
import type { DailyTrendPoint } from './dailyTrendPoint';
import type { DepartmentPerformance } from './departmentPerformance';

export interface AnalyticsData {
  categoryBreakdown: ChartDataPoint[];
  dailyTrend: DailyTrendPoint[];
  priorityDistribution: ChartDataPoint[];
  wardBreakdown: ChartDataPoint[];
  departmentPerformance: DepartmentPerformance[];
}