# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CivicPulse AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  title: string;
  description: string;
  location: string;
}